rename src and main to make library import more meaningful
