+++ wrf_analysis_toolkit/SkewT.py
@@ -5,7 +5,7 @@ from wrf import ll_to_xy, getvar
 import matplotlib.pyplot as plt
 import numpy as np
 
-import src.SensibleVariables as sv
+import wrf_analysis_toolkit.SensibleVariables as sv
 
 
 def Plot_SkewT(ncfile, ti, svariable, outfname="MyPlot.png", save_pdf=0):
