+++ wrf_analysis_toolkit_cli.py
@@ -4,12 +4,12 @@ import sys
 # Adds folder to python path search
 sys.path.insert(1, "/".join(__file__.split("/")[:-2]))
 
-from src.Animate import Animate
-from src.TerrainPlots import Terrain
-from src.CSV_Data import CSV_Data
-import src.SensibleVariables as sv
-from src.MP4Compare import *
-from src.WRFCompare import *
+from wrf_analysis_toolkit.Animate import Animate
+from wrf_analysis_toolkit.TerrainPlots import Terrain
+from wrf_analysis_toolkit.CSV_Data import CSV_Data
+import wrf_analysis_toolkit.SensibleVariables as sv
+from wrf_analysis_toolkit.MP4Compare import *
+from wrf_analysis_toolkit.WRFCompare import *
 
 import argparse
